+++ include/Nazara/Renderer/RenderTexture.hpp
@@ -43,10 +43,8 @@ namespace Nz
 
 			void Detach(AttachmentPoint attachmentPoint, UInt8 index);
 
-			unsigned int GetHeight() const override;
 			RenderTargetParameters GetParameters() const override;
 			Vector2ui GetSize() const;
-			unsigned int GetWidth() const override;
 
 			bool IsComplete() const;
 			bool IsRenderable() const override;