+++ ChangeLog.md
@@ -4,6 +4,8 @@ Nazara Engine:
 - VertexMapper:GetComponentPtr no longer throw an error if component is disabled or incompatible with template type, instead a null pointer is returned.
 - Bitset swap operation is now correctly marked as noexcept`
 - Mesh loaders now takes MeshParams vertexDeclaration into account
+- ⚠️ Replaced RenderTarget::Get[Height|Width] by RenderTarget::GetSize
+- ⚠️ Removed Window::Get[Height|Width] methods
 
 # 0.4: